Only Install Steam If Needed
To allow this script to be run twice, only use apt update and apt
install for steam and other components at the first run.

Please note that this will not cover if a users installed steam manually
but did not install steam-devices or x11-utils. Skipping this case here
since we assume that this script is run on vanilla ubuntu.